grpc_api/pyproject.toml
[project]
name = "grpc-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "orjson>=3.10.7",
    "pandas>=2.2.3",
    "protobuf>=5.28.2",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "mypy-protobuf>=3.6.0",
    "httpx>=0.27.2",
    "pyarrow>=18.0.0",
    "watchdog[watchmedo]>=6.0.0",
]